// frontend/src/model/apiResponseUserResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Bus Ticket Booking API
 * REST API for searching trips, booking seats and managing bus ticket operations.
 * OpenAPI spec version: v1.0
 */
import type { UserResponse } from './userResponse';

export interface ApiResponseUserResponse {
  status?: number;
  message?: string;
  data?: UserResponse;
}